+++ src/main/scala/ie/tcd/slscs/itut/DictionaryConverter/TrxProc.scala
@@ -154,76 +154,6 @@ object TrxProc {
   // TODO: mutable rule? maybe divide by contents - add macros separately, e.g.
   // TODO: convert macros from SimpleTextMacro
 
-  trait MacroAttr
-  case class LemmaMacroAttr(s: String) extends MacroAttr
-  case class ListMacroAttr(s: String) extends MacroAttr
-  case class KVMacroAttr(k: String, v: String) extends MacroAttr
-  case class KeyOnlyMacroAttr(k: String) extends MacroAttr
-  def convertSimpleTextMacroAttr(in: SimpleTextMacroAttr): MacroAttr = {
-    if(in.getKey == "lemma") {
-      LemmaMacroAttr(in.getValue)
-    } else if(in.getKey == "inlist") {
-        ListMacroAttr(in.getValue)
-    } else if(in.getValue == "" || in.getValue == null) {
-      KeyOnlyMacroAttr(in.getKey)
-    } else {
-      KVMacroAttr(in.getKey, in.getValue)
-    }
-  }
-  // TODO: kv -> assign to variable, or clip
-  // a map with valid attributes for that pos is required, should be passed from elsewhere
-  def convertMacroAttrToLet(in: MacroAttr, pos: Int): LetElement = in match {
-    case LemmaMacroAttr(s) => LetElement(ClipElement(pos.toString, "tl", "lem", null, null, null), LitElement(s))
-    case KVMacroAttr(k, v) => LetElement(ClipElement(pos.toString, "tl", k, null, null, null), LitTagElement(v))
-    case _ => throw new Exception("Can't convert this tag")
-  }
-  def convertMacroAttrToTest(in: MacroAttr, pos: Int): TestElement = in match {
-    case LemmaMacroAttr(s) => TestElement(null, EqualElement(true, List[ValueElement](ClipElement(pos.toString, "tl", "lem", null, null, null), LitElement(s))))
-    case ListMacroAttr(s) => TestElement(null, InElement(ClipElement(pos.toString, "tl", "lem", null, null, null), ListElement(s), true))
-    case KVMacroAttr(k, v) => TestElement(null, EqualElement(true, List[ValueElement](ClipElement(pos.toString, "tl", k, null, null, null), LitTagElement(v))))
-    case _ => throw new Exception("Can't convert this tag")
-  }
-  abstract class BaseTextMacroEntry
-  abstract class SrcTextMacroEntry extends BaseTextMacroEntry
-  abstract class TrgTextMacroEntry extends BaseTextMacroEntry
-  case class SrcBaseMacroEntry(pos: Int, matches: List[MacroAttr]) extends SrcTextMacroEntry
-  case class SrcInsertionMacroEntry(pos: Int, matches: List[MacroAttr]) extends SrcTextMacroEntry
-  case class SrcDeletionMacroEntry(pos: Int, matches: List[MacroAttr]) extends SrcTextMacroEntry
-  case class SrcChunkMacroEntry(pos: Int, matches: List[MacroAttr]) extends SrcTextMacroEntry
-  case class TrgBaseMacroEntry(pos: Int, matches: List[MacroAttr], targets: List[SrcTextMacroEntry]) extends TrgTextMacroEntry
-  case class TrgInsertionMacroEntry(pos: Int, matches: List[MacroAttr], targets: List[SrcTextMacroEntry]) extends TrgTextMacroEntry
-  case class TrgDeletionMacroEntry(pos: Int, matches: List[MacroAttr], targets: List[SrcTextMacroEntry]) extends TrgTextMacroEntry
-  case class TrgChunkMacroEntry(pos: Int, matches: List[MacroAttr], targets: List[SrcTextMacroEntry]) extends TrgTextMacroEntry
-  case class SimpleTextMacro(name: String, appliesTo: List[String], entries: List[TrgTextMacroEntry])
-  def convertTextMacroEntry(in: JSTMEntry): BaseTextMacroEntry = {
-    if(in.hasTarget) {
-      convertTrgTextMacroEntry(in)
-    } else {
-      convertSrcTextMacroEntry(in)
-    }
-  }
-  def convertSrcTextMacroEntry(in: JSTMEntry): SrcTextMacroEntry = {
-    if(in.isInsertion) {
-      SrcInsertionMacroEntry(in.getPosition, in.getAttrs.asScala.map{convertSimpleTextMacroAttr}.toList)
-    } else if(in.isDeletion) {
-      SrcDeletionMacroEntry(in.getPosition, in.getAttrs.asScala.map{convertSimpleTextMacroAttr}.toList)
-    } else if(in.isChunkAlignment) {
-      SrcChunkMacroEntry(in.getPosition, in.getAttrs.asScala.map{convertSimpleTextMacroAttr}.toList)
-    } else {
-      SrcBaseMacroEntry(in.getPosition, in.getAttrs.asScala.map{convertSimpleTextMacroAttr}.toList)
-    }
-  }
-  def convertTrgTextMacroEntry(in: JSTMEntry): TrgTextMacroEntry = {
-    if(in.isInsertion) {
-      TrgInsertionMacroEntry(in.getPosition, in.getAttrs.asScala.map{convertSimpleTextMacroAttr}.toList, in.getTarget.asScala.map{convertSrcTextMacroEntry}.toList)
-    } else if(in.isDeletion) {
-      TrgDeletionMacroEntry(in.getPosition, in.getAttrs.asScala.map{convertSimpleTextMacroAttr}.toList, in.getTarget.asScala.map{convertSrcTextMacroEntry}.toList)
-    } else if(in.isChunkAlignment) {
-      TrgChunkMacroEntry(in.getPosition, in.getAttrs.asScala.map{convertSimpleTextMacroAttr}.toList, in.getTarget.asScala.map{convertSrcTextMacroEntry}.toList)
-    } else {
-      TrgBaseMacroEntry(in.getPosition, in.getAttrs.asScala.map{convertSimpleTextMacroAttr}.toList, in.getTarget.asScala.map{convertSrcTextMacroEntry}.toList)
-    }
-  }
 
   case class RuleMetadata(ruleid: String, rulecomment: String)
